Estimate heritability

ldsc_h2() uses ldscore regression to estimate the heritability of a trait from GWAS summary statistics and reference LD information.

Usage

ldsc_h2(
  munged_sumstats,
  ancestry,
  sample_prev = NA,
  population_prev = NA,
  ld,
  wld,
  n_blocks = 200,
  chisq_max = NA
)

Arguments

munged_sumstats

Either a dataframe, or a path to a file containing munged summary statistics. Must contain at least columns named SNP (rsid), A1 (effect allele), A2 (non-effect allele), N (total sample size) and Z (Z-score)

ancestry

(character) One of "AFR", "AMR", "CSA", "EAS", "EUR", or "MID", which will utilize the appropriate built-in ld and wld files from Pan-UK Biobank. If empty or NULL, the user must specify paths to ld and wld files.

sample_prev

(numeric) For binary traits, this should be the prevalence of cases in the current sample, used for conversion from observed heritability to liability-scale heritability. The default is NA, which is appropriate for quantitative traits or estimating heritability on the observed scale.

population_prev

(numeric) For binary traits, this should be the population prevalence of the trait, used for conversion from observed heritability to liability-scale heritability. The default is NA, which is appropriate for quantitative traits or estimating heritability on the observed scale.

ld

(character) Path to directory containing ld score files, ending in *.l2.ldscore.gz. Default is NA, which will utilize the built-in ld score files from Pan-UK Biobank for the ancestry specified in ancestry.

wld

(character) Path to directory containing weight files. Default is NA, which will utilize the built-in weight files from Pan-UK Biobank for the ancestry specified in ancestry.

n_blocks

(numeric) Number of blocks used to produce block jackknife standard errors. Default is 200

chisq_max

(numeric) Maximum value of Z^2 for SNPs to be included in LD-score regression. Default is to set chisq_max to the maximum of 80 and N*0.001.

Value

A tibble containing heritability information. If sample_prev and population_prev were provided, the heritability estimate will also be returned on the liability scale.

Examples

# \donttest{
ldsc_h2(sumstats_munged_example(example = "BMI", dataframe = TRUE), ancestry = "EUR")
#> ℹ Using EUR reference from Pan-UKB
#> ✔ Using EUR reference from Pan-UKB [26ms]
#> 
#> ℹ Reading LD Scores
#> ✔ Reading LD Scores [489ms]
#> 
#> ℹ Reading weights
#> ✔ Reading weights [380ms]
#> 
#> ℹ Reading M
#> ✔ Reading M [25ms]
#> 
#> ℹ Reading summary statistics
#> ✔ Reading summary statistics [356ms]
#> 
#> ℹ Reading summary statistics from dataframe
#> ✔ Reading summary statistics from dataframe [18ms]
#> 
#> ℹ Merging summary statistics with LD-score files
#> ℹ 1075707/1168351 SNPs remain after merging with LD-score files
#> ℹ Merging summary statistics with LD-score files
#> ℹ Removed 20 SNPs with Chi^2 > 336.107; 1075687 SNPs remain
#> ℹ Merging summary statistics with LD-score files
#> ✔ Merging summary statistics with LD-score files [10.6s]
#> 
#> ℹ Estimating heritability
#> ✔ Estimating heritability [508ms]
#> 
#> # A tibble: 1 × 10
#>   mean_…¹ lambd…² inter…³ inter…⁴  ratio ratio…⁵ h2_ob…⁶ h2_ob…⁷  h2_Z      h2_p
#>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl> <dbl>     <dbl>
#> 1    2.72    2.13    1.11  0.0211 0.0631  0.0123   0.238 0.00814  29.2 1.28e-187
#> # … with abbreviated variable names ¹​mean_chisq, ²​lambda_gc, ³​intercept,
#> #   ⁴​intercept_se, ⁵​ratio_se, ⁶​h2_observed, ⁷​h2_observed_se
# }