Estimate cross-trait genetic correlations

ldsc_rg() uses ldscore regression to estimate the pairwise genetic correlations between traits. The function relies on named lists of traits, sample prevalences, and populatin prevalences. The name of each trait should be consistent across each argument.

Usage

ldsc_rg(
  munged_sumstats,
  ancestry,
  sample_prev = NA,
  population_prev = NA,
  ld,
  wld,
  n_blocks = 200,
  chisq_max = NA
)

Arguments

munged_sumstats

(list) A named list of dataframes, or paths to files containing munged summary statistics. Each set of munged summary statistics contain at least columns named SNP (rsid), A1 (effect allele), A2 (non-effect allele), N (total sample size) and Z (Z-score)

ancestry

(character) One of "AFR", "AMR", "CSA", "EAS", "EUR", or "MID", which will utilize the appropriate built-in ld and wld files from Pan-UK Biobank. If empty or NULL, the user must specify paths to ld and wld files.

sample_prev

(list) A named list containing the prevalence of cases in the current sample, used for conversion from observed heritability to liability-scale heritability. The default is NA, which is appropriate for quantitative traits or estimating heritability on the observed scale.

population_prev

(list) A named list containing the population prevalence of the trait, used for conversion from observed heritability to liability-scale heritability. The default is NA, which is appropriate for quantitative traits or estimating heritability on the observed scale.

ld

(character) Path to directory containing ld score files, ending in *.l2.ldscore.gz. Default is NA, which will utilize the built-in ld score files from Pan-UK Biobank for the ancestry specified in ancestry.

wld

(character) Path to directory containing weight files. Default is NA, which will utilize the built-in weight files from Pan-UK Biobank for the ancestry specified in ancestry.

n_blocks

(numeric) Number of blocks used to produce block jackknife standard errors. Default is 200

chisq_max

(numeric) Maximum value of Z^2 for SNPs to be included in LD-score regression. Default is to set chisq_max to the maximum of 80 and N*0.001.

Value

A list of class ldscr_list containing heritablilty and genetic correlation information

h2 = tibble containing heritability information for each trait. If sample_prev and population_prev were provided, the heritability estimates will also be returned on the liability scale.

rg = tibble containing pairwise genetic correlations information.

raw = A list of correlation/covariance matrices

Details

This function estimates the pairwise genetic correlations between an arbitrary number of traits. The function also estimates heritability for each individual trait. There is a ggplot2::autoplot() method for visualizing a heatmap of the results.

Examples

# \donttest{
ldsc_rg(munged_sumstats = list("APOB" = sumstats_munged_example(example = "APOB"), "LDL" = sumstats_munged_example(example = "LDL")), ancestry = "EUR")
#> ℹ Using EUR reference from Pan-UKB
#> ℹ No sample prevalence data provided. Estimating heritabilities on the observed scale.
#> ℹ Using EUR reference from Pan-UKB
#> ✔ Using EUR reference from Pan-UKB [550ms]
#> 
#> ℹ Reading LD Scores
#> ✔ Reading LD Scores [364ms]
#> 
#> ℹ Reading weights
#> ✔ Reading weights [368ms]
#> 
#> ℹ Reading M
#> ✔ Reading M [21ms]
#> 
#> ℹ Reading summary statistics
#> ℹ Reading summary statistics for 'APOB' from dataframe
#> ✔ Reading summary statistics for 'APOB' from dataframe [8ms]
#> 
#> ℹ Reading summary statistics
#> ℹ Merging 'APOB' with LD-score files
#> ℹ 1087712/1089172 SNPs remain after merging 'APOB' with LD-score files
#> ℹ Merging 'APOB' with LD-score files
#> ℹ Removed 171 SNPs with Chi^2 > 342.59 from 'APOB'; 1087541 SNPs remain
#> ℹ Merging 'APOB' with LD-score files
#> ✔ Merging 'APOB' with LD-score files [4.9s]
#> 
#> ℹ Reading summary statistics
#> ℹ Reading summary statistics for 'LDL' from dataframe
#> ✔ Reading summary statistics for 'LDL' from dataframe [8ms]
#> 
#> ℹ Reading summary statistics
#> ℹ Merging 'LDL' with LD-score files
#> ℹ 1087712/1089172 SNPs remain after merging 'LDL' with LD-score files
#> ℹ Merging 'LDL' with LD-score files
#> ℹ Removed 127 SNPs with Chi^2 > 343.621 from 'LDL'; 1087585 SNPs remain
#> ℹ Merging 'LDL' with LD-score files
#> ✔ Merging 'LDL' with LD-score files [4.2s]
#> 
#> ℹ Reading summary statistics
#> ✔ Reading summary statistics [9.2s]
#> 
#> ℹ Estimating heritability for 'APOB'
#> ✔ Estimating heritability for 'APOB' [493ms]
#> 
#> ℹ Estimating genetic covariance for for 'APOB' and 'LDL'
#> ✔ Estimating genetic covariance for for 'APOB' and 'LDL' [2.5s]
#> 
#> ℹ Estimating heritability for 'LDL'
#> ✔ Estimating heritability for 'LDL' [208ms]
#> 
#> $h2
#> # A tibble: 2 × 11
#>   trait mean_chisq lambda_gc inter…¹ inter…² ratio ratio…³ h2_ob…⁴ h2_ob…⁵  h2_Z
#>   <chr>      <dbl>     <dbl>   <dbl>   <dbl> <dbl>   <dbl>   <dbl>   <dbl> <dbl>
#> 1 APOB        1.72      1.32    1.13  0.0398 0.173  0.0551  0.0853 0.00930  9.18
#> 2 LDL         1.67      1.29    1.10  0.0382 0.156  0.0571  0.0792 0.00865  9.16
#> # … with 1 more variable: h2_p <dbl>, and abbreviated variable names
#> #   ¹​intercept, ²​intercept_se, ³​ratio_se, ⁴​h2_observed, ⁵​h2_observed_se
#> 
#> $rg
#> # A tibble: 1 × 5
#>   trait1 trait2    rg rg_se     rg_p
#>   <chr>  <chr>  <dbl> <dbl>    <dbl>
#> 1 APOB   LDL    0.936 0.102 5.88e-20
#> 
#> $raw
#> $raw$V
#>              [,1]         [,2]         [,3]
#> [1,] 8.641845e-05 7.696781e-05 7.616337e-05
#> [2,] 7.696781e-05 7.077810e-05 7.169640e-05
#> [3,] 7.616337e-05 7.169640e-05 7.478091e-05
#> 
#> $raw$S
#>            APOB        LDL
#> APOB 0.08531389 0.07695010
#> LDL  0.07695010 0.07919108
#> 
#> $raw$I
#>          [,1]     [,2]
#> [1,] 1.125266 1.047318
#> [2,] 1.047318 1.104392
#> 
#> $raw$N
#>        [,1]     [,2]   [,3]
#> [1,] 342590 343105.1 343621
#> 
#> $raw$m
#> [1] 6805960
#> 
#> $raw$V_Stand
#>            [,1]       [,2]       [,3]
#> [1,] 0.01187318 0.01097593 0.01127327
#> [2,] 0.01097593 0.01047618 0.01101471
#> [3,] 0.01127327 0.01101471 0.01192445
#> 
#> $raw$S_Stand
#>           APOB       LDL
#> APOB 1.0000000 0.9361839
#> LDL  0.9361839 1.0000000
#> 
#> $raw$SE_Stand
#>           APOB       LDL
#> APOB 0.1089641 0.0000000
#> LDL  0.1023532 0.1091991
#> 
#> 
#> attr(,"class")
#> [1] "ldscr_list" "list"      
# }